+++ tech1-framework-b2b-mongodb-security-jwt/src/main/java/io/tech1/framework/b2b/mongodb/security/jwt/handlers/exceptions/JwtAuthenticationEntryPointExceptionHandler.java
@@ -57,13 +57,13 @@ public class JwtAuthenticationEntryPointExceptionHandler implements Authenticati
                     )
             );
             this.securityJwtIncidentPublisher.publishAuthenticationLoginFailureUsernamePassword(
-                    IncidentAuthenticationLoginFailureUsernamePassword.of(
+                    new IncidentAuthenticationLoginFailureUsernamePassword(
                             username,
                             password
                     )
             );
             this.securityJwtIncidentPublisher.publishAuthenticationLoginFailureUsernameMaskedPassword(
-                    IncidentAuthenticationLoginFailureUsernameMaskedPassword.of(
+                    new IncidentAuthenticationLoginFailureUsernameMaskedPassword(
                             username,
                             password
                     )