+++ tech1-framework-b2b-mongodb-security-jwt/src/test/java/io/tech1/framework/b2b/mongodb/security/jwt/events/subscribers/base/BaseSecurityJwtSubscriberTest.java
@@ -212,7 +212,7 @@ public class BaseSecurityJwtSubscriberTest {
         // Assert
         verify(this.userSessionService).saveUserRequestMetadata(eq(event));
         verify(this.userEmailService).executeAuthenticationLogin(eq(new FunctionAuthenticationLoginEmail(event.getUsername(), event.getEmail(), userSession.getRequestMetadata())));
-        verify(this.securityJwtIncidentPublisher).publishAuthenticationLogin(eq(IncidentAuthenticationLogin.of(event.getUsername(), userSession.getRequestMetadata())));
+        verify(this.securityJwtIncidentPublisher).publishAuthenticationLogin(eq(new IncidentAuthenticationLogin(event.getUsername(), userSession.getRequestMetadata())));
     }
 
     @Test
@@ -236,6 +236,6 @@ public class BaseSecurityJwtSubscriberTest {
         // Assert
         verify(this.userSessionService).saveUserRequestMetadata(eq(event));
         verify(this.userEmailService).executeSessionRefreshed(eq(new FunctionSessionRefreshedEmail(event.getUsername(), event.getEmail(), userSession.getRequestMetadata())));
-        verify(this.securityJwtIncidentPublisher).publishSessionRefreshed(eq(IncidentSessionRefreshed.of(event.getUsername(), userSession.getRequestMetadata())));
+        verify(this.securityJwtIncidentPublisher).publishSessionRefreshed(eq(new IncidentSessionRefreshed(event.getUsername(), userSession.getRequestMetadata())));
     }
 }