+++ tech1-framework-b2b-mongodb-security-jwt/src/test/java/io/tech1/framework/b2b/mongodb/security/jwt/handlers/exceptions/JwtAuthenticationEntryPointExceptionHandlerTest.java
@@ -162,13 +162,13 @@ public class JwtAuthenticationEntryPointExceptionHandlerTest {
                 )
         ));
         verify(this.securityJwtIncidentPublisher).publishAuthenticationLoginFailureUsernamePassword(eq(
-                IncidentAuthenticationLoginFailureUsernamePassword.of(
+                new IncidentAuthenticationLoginFailureUsernamePassword(
                         username,
                         password
                 )
         ));
         verify(this.securityJwtIncidentPublisher).publishAuthenticationLoginFailureUsernameMaskedPassword(eq(
-                IncidentAuthenticationLoginFailureUsernameMaskedPassword.of(
+                new IncidentAuthenticationLoginFailureUsernameMaskedPassword(
                         username,
                         password
                 )