+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/domain/registration/IncidentRegistration1.java
@@ -4,7 +4,7 @@ import io.tech1.framework.domain.base.Username;
 import lombok.Data;
 
 // Lombok
-@Data(staticConstructor = "of")
+@Data
 public class IncidentRegistration1 {
     private final Username username;
 }