+++ tech1-framework-incidents/src/test/java/io/tech1/framework/incidents/domain/authetication/IncidentAuthenticationLoginFailureUsernameMaskedPasswordTest.java
@@ -29,7 +29,7 @@ public class IncidentAuthenticationLoginFailureUsernameMaskedPasswordTest {
 
 
         // Act
-        var actual = IncidentAuthenticationLoginFailureUsernameMaskedPassword.of(
+        var actual = new IncidentAuthenticationLoginFailureUsernameMaskedPassword(
                 username,
                 Password.of(password)
         );